Recreated file structure and rebuilt project
