+++ .gitignore
@@ -37,3 +37,4 @@ nosetests.xml
 
 # Virtual environment
 venv
+.cache
