--- app/static/styles/recipe.css
@@ -1,61 +1,0 @@
-*{
-  margin:0;
-  padding:0;
-  }
-
-body{
-    font-family:tahoma;
-    background-image:linear-gradient(rgba(0,0,0,0.7),rgba(0,0,0,0.7)),url("img/cover.jpg");
-    height:100vh;
-    background-position:center;
-    background-size:cover;
-    }
-
-.hero{
-    position:absolute;
-    width:1140px;
-    left:50%;
-    top:50%;
-    transform: translate(-50%,-50%);
-
-    }
-
-header h1{
-    color:#ffffff;
-    }
-
-.row{
-    max-width:1140px;
-    margin:0 auto;
-    }
-
-.main-nav{
-    float:right;
-    list-style:none;
-    }
-
-.main-nav li{
-    display:inline-block;
-    margin-left:20px;
-    margin-top:40px;
-
-    }
-
-.main-nav li a{
-    color:#ffffff;
-    text-decoration:none;
-    font-weight:bold;
-    }
-
- .main-nav li a:hover{
-    color:#13d1c5;
-    border-bottom:2px solid #13d1c5;
-    transition:all 0.5s ease-in;
-    padding:10px;
-
-    }
-.logo img{
-    height:100px;
-    width:100px;
-    linear-gradient(rgba(0,0,0,0.7),rgba(0,0,0,0.7));
-    }
