--- config.py
@@ -1,2 +1,0 @@
-#enable debug feature in flask app. This should be false in production
-Debug = True
