+++ templates/Categories.html
@@ -12,11 +12,11 @@
         <div class="row">
             <nav>
                 <ul class="main-nav">
-                    <li><a href="index.html">Meet us</a></li>
-                    <li><a href="myrecipes.html">My recipes</a></li>
-                    <li><a href="Sign-in.html" >Login</a></li>
-                    <li><a href="Sign-up.html">Sign up</a></li>
-                    <li><a href="Categories.html" id="current">Categories</a></li>
+                    <li><a href="/">Meet us</a></li>
+                    <li><a href="/myrecipes">My recipes</a></li>
+                    <li><a href="/login" >Login</a></li>
+                    <li><a href="/Sign-up">Sign up</a></li>
+                    <li><a href="/Categories" id="current">Categories</a></li>
                 </ul>
             </nav>
         </div>
