+++ templates/myrecipes.html
@@ -23,7 +23,7 @@
         </p>
         <button class="button" type="submit">Edit</button>
         <button class="button" type="submit">Delete</button>
-        <button class="button" onclick="location.href='index.html'" type="submit">Home</button>
+        <button class="button" onclick="location.href='/'" type="submit">Home</button>
     </section>
     <div class="pimg2">
         <div class="ptext">
@@ -42,7 +42,7 @@
         </p>
         <button class="button" type="submit">Edit</button>
         <button class="button" type="submit">Delete</button>
-        <button class="button" onclick="location.href='index.html'" type="submit">Home</button>
+        <button class="button" onclick="location.href='/'" type="submit">Home</button>
     </section>
 
 
@@ -63,7 +63,7 @@
         </p>
         <button class="button" type="submit">Edit</button>
         <button class="button" type="submit">Delete</button>
-        <button class="button" onclick="location.href='index.html'" type="submit">Home</button>
+        <button class="button" onclick="location.href='/'" type="submit">Home</button>
     </section>
     <div class="pimg1">
         <div class="ptext">
